add rest but jms queue does not work
