+++ MainSystemImpl/src/main/java/rest/AlertService.java
@@ -2,19 +2,13 @@ package rest;
 
 import database.ParkingSpotDAO;
 import entities.ParkingSpot;
-import events.EventDetectionManager;
-import org.json.simple.JSONObject;
-import org.json.simple.parser.JSONParser;
-import org.json.simple.parser.ParseException;
+import nevents.EventDetectionManager;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 
 @Path("/alerts")
 public class AlertService {
